+++ yarn-project/acir-simulator/src/client/db_oracle.ts
@@ -2,8 +2,9 @@ import { AztecAddress } from '@aztec/foundation/aztec-address';
 import { EthAddress } from '@aztec/foundation/eth-address';
 import { Fr, Point } from '@aztec/foundation/fields';
 import { FunctionAbi } from '@aztec/foundation/abi';
+import { PartialContractAddress } from '@aztec/circuits.js';
+
 import { CommitmentsDB } from '../index.js';
-import { PartialContractAddress } from '@aztec/types';
 
 /**
  * The format that noir contracts use to get notes.
