+++ yarn-project/aztec-rpc/src/aztec_rpc_server/aztec_rpc_server.ts
@@ -1,5 +1,5 @@
 import { AztecNode } from '@aztec/aztec-node';
-import { AztecAddress, EthAddress, FunctionData } from '@aztec/circuits.js';
+import { AztecAddress, EthAddress, FunctionData, PartialContractAddress } from '@aztec/circuits.js';
 import { ContractAbi, FunctionType, encodeArguments, generateFunctionSelector } from '@aztec/foundation/abi';
 import { Fr, Point } from '@aztec/foundation/fields';
 import { createDebugLogger } from '@aztec/foundation/log';
@@ -15,7 +15,6 @@ import {
   L2BlockL2Logs,
   LogType,
   NodeInfo,
-  PartialContractAddress,
   PublicKey,
   Tx,
   TxExecutionRequest,
