+++ yarn-project/aztec-rpc/src/database/database.ts
@@ -1,7 +1,8 @@
-import { PartialContractAddress, ContractDatabase, TxHash, PublicKey } from '@aztec/types';
+import { ContractDatabase, TxHash, PublicKey } from '@aztec/types';
 import { AztecAddress } from '@aztec/foundation/aztec-address';
 import { Fr, Point } from '@aztec/foundation/fields';
 import { MerkleTreeId } from '@aztec/types';
+import { PartialContractAddress } from '@aztec/circuits.js';
 
 import { NoteSpendingInfoDao } from './note_spending_info_dao.js';
 import { TxDao } from './tx_dao.js';
