+++ yarn-project/aztec-rpc/src/database/memory_db.ts
@@ -1,7 +1,8 @@
-import { PartialContractAddress, TxHash } from '@aztec/types';
+import { TxHash } from '@aztec/types';
 import { AztecAddress } from '@aztec/foundation/aztec-address';
 import { Fr, Point } from '@aztec/foundation/fields';
 import { MerkleTreeId, PublicKey } from '@aztec/types';
+import { PartialContractAddress } from '@aztec/circuits.js';
 
 import { MemoryContractDatabase } from '../contract_database/index.js';
 import { Database, GetOptions } from './database.js';
