+++ yarn-project/aztec-rpc/src/index.ts
@@ -2,9 +2,9 @@ export * from './aztec_rpc_server/index.js';
 export * from './aztec_rpc_http/index.js';
 export * from './crypto/types.js';
 
-export { Tx, TxHash, PartialContractAddress } from '@aztec/types';
+export { Tx, TxHash } from '@aztec/types';
 
-export { TxRequest, CircuitsWasm } from '@aztec/circuits.js';
+export { TxRequest, CircuitsWasm, PartialContractAddress } from '@aztec/circuits.js';
 export * from '@aztec/foundation/fields';
 export * from '@aztec/foundation/eth-address';
 export * from '@aztec/foundation/aztec-address';
