+++ yarn-project/aztec.js/src/contract_deployer/deploy_method.ts
@@ -1,17 +1,15 @@
-import { CircuitsWasm, ContractDeploymentData, TxContext, getContractDeploymentInfo } from '@aztec/circuits.js';
+import {
+  CircuitsWasm,
+  ContractDeploymentData,
+  PartialContractAddress,
+  TxContext,
+  getContractDeploymentInfo,
+} from '@aztec/circuits.js';
 import { ContractAbi } from '@aztec/foundation/abi';
 import { AztecAddress } from '@aztec/foundation/aztec-address';
 import { EthAddress } from '@aztec/foundation/eth-address';
 import { Fr } from '@aztec/foundation/fields';
-import {
-  AztecRPC,
-  ExecutionRequest,
-  PackedArguments,
-  PartialContractAddress,
-  PublicKey,
-  Tx,
-  TxExecutionRequest,
-} from '@aztec/types';
+import { AztecRPC, ExecutionRequest, PackedArguments, PublicKey, Tx, TxExecutionRequest } from '@aztec/types';
 
 import { BaseWallet, Wallet } from '../aztec_rpc_client/wallet.js';
 import { Contract, ContractFunctionInteraction, SendMethodOptions } from '../contract/index.js';
