+++ yarn-project/circuits.js/package.json
@@ -38,7 +38,6 @@
   },
   "dependencies": {
     "@aztec/foundation": "workspace:^",
-    "@aztec/types": "workspace:^",
     "@msgpack/msgpack": "^3.0.0-beta2",
     "@noble/curves": "^1.0.0",
     "@types/lodash.camelcase": "^4.3.7",
