+++ yarn-project/circuits.js/src/index.ts
@@ -4,3 +4,4 @@ export * from './kernel/index.js';
 export * from './rollup/index.js';
 export * from './utils/jsUtils.js';
 export * from './contract/index.js';
+export * from './types/index.js';
