+++ yarn-project/circuits.js/tsconfig.json
@@ -8,9 +8,6 @@
   "references": [
     {
       "path": "../foundation"
-    },
-    {
-      "path": "../types"
     }
   ],
   "include": ["src"]
