+++ yarn-project/types/src/keys/key_store.ts
@@ -1,7 +1,6 @@
-import { Point } from '@aztec/circuits.js';
+import { PublicKey } from '@aztec/circuits.js';
 
-/** Represents a user public key. */
-export type PublicKey = Point;
+export { PublicKey } from '@aztec/circuits.js';
 
 /**
  * Represents a secure storage for managing keys.
