--- yarn-project/types/src/partial_contract_address.ts
@@ -1,6 +1,0 @@
-import { Fr } from '@aztec/circuits.js';
-
-/**
- * A type to represent a partially constructed contract address.
- */
-export type PartialContractAddress = Fr;
