+++ yarn-project/types/src/tx/tx.ts
@@ -3,6 +3,7 @@ import {
   Fr,
   MAX_PUBLIC_CALL_STACK_LENGTH_PER_TX,
   KernelCircuitPublicInputs,
+  PartialContractAddress,
   Proof,
   PublicCallRequest,
 } from '@aztec/circuits.js';
@@ -12,7 +13,6 @@ import { BufferReader, numToUInt32BE } from '@aztec/foundation/serialize';
 
 import { EncodedContractFunction } from '../contract_data.js';
 import { TxL2Logs } from '../logs/tx_l2_logs.js';
-import { PartialContractAddress } from '../partial_contract_address.js';
 import { TxHash } from './tx_hash.js';
 
 /**
